requirements.txt
websockets
requests
webexteamssdk
PyYAML  >= 5.1
pymysql==0.7.11
mysql-connector
numpy
regex
configparser
pymongo
